feat: make hcn_new trainable and inferable after dev merge

+++ deeppavlov/run_model.py
@@ -23,12 +23,12 @@ from deeppavlov.core.commands.utils import set_usr_dir
 usr_dir = None
 
 try:
-    MODEL_CONFIG_PATH = 'models/classifiers/intents/config.json'
+    MODEL_CONFIG_PATH = 'skills/hcn_new/config.json'
     usr_dir = set_usr_dir(MODEL_CONFIG_PATH)
     train_model_from_config(MODEL_CONFIG_PATH)
     interact_model(MODEL_CONFIG_PATH)
 except Exception:
     raise
 finally:
-    if usr_dir:
+    if usr_dir and not usr_dir.rglob('*'):
         usr_dir.rmdir()
